+++ app/(protected)/layout.tsx
@@ -1,6 +1,20 @@
 import React from 'react'
 import type { Metadata } from 'next'
-import { AIPopup, Navbar } from '@/components'
+import { AIPopup, Navbar, NotificationSnackBar } from '@/components'
+
+import localFont from 'next/font/local'
+import './globals.css'
+
+const geistSans = localFont({
+  src: './fonts/GeistVF.woff',
+  variable: '--font-geist-sans',
+  weight: '100 900'
+})
+const geistMono = localFont({
+  src: './fonts/GeistMonoVF.woff',
+  variable: '--font-geist-mono',
+  weight: '100 900'
+})
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -14,7 +28,8 @@ export default function RootLayout({
 }>) {
   return (
     <html lang='en'>
-      <body className={`antialiased bg-background`}>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased bg-background`}>
+        <NotificationSnackBar />
         <Navbar />
         {children}
         <AIPopup />
